rename icon test helpers file
