+++ src/element-testing/fixture-test.ts
@@ -1,8 +1,28 @@
 import {fixtureCleanup} from '@open-wc/testing';
+import {isPromiseLike} from 'augment-vir';
 
-export function fixtureTest(callback: () => Promise<void>) {
+export function createFixtureTest(callback: () => Promise<void>) {
     return async () => {
-        await callback();
-        fixtureCleanup();
+        await withFixtureCleanup(callback);
     };
+}
+
+export function withFixtureCleanup<T>(callback: () => T): T;
+export function withFixtureCleanup<T>(callback: () => Promise<T>): Promise<T>;
+export function withFixtureCleanup<T>(callback: () => Promise<T> | T): T | Promise<T> {
+    const result = callback();
+    if (isPromiseLike(result)) {
+        return new Promise<T>(async (resolve, reject) => {
+            try {
+                const awaitedResult = await result;
+                fixtureCleanup();
+                resolve(awaitedResult);
+            } catch (error) {
+                reject(error);
+            }
+        });
+    } else {
+        fixtureCleanup();
+        return result;
+    }
 }
