+++ src/element-testing/icon-test-helpers.ts
