+++ src/elements/toniq-checkbox/toniq-checkbox.test.ts
@@ -1,6 +1,6 @@
 import {assert, fixture} from '@open-wc/testing';
 import {assign, html} from 'element-vir';
-import {fixtureTest} from '../../element-testing/fixture-test';
+import {createFixtureTest} from '../../element-testing/fixture-test';
 import {getTextContentThroughShadow} from '../../element-testing/query-through-shadow';
 import {ToniqCheckbox} from './toniq-checkbox.element';
 
@@ -12,7 +12,7 @@ describe(ToniqCheckbox.tagName, () => {
 
     it(
         'should render assigned text',
-        fixtureTest(async () => {
+        createFixtureTest(async () => {
             const textToRender = 'Orange';
 
             const rendered = await fixture(
