+++ src/elements/toniq-chip/toniq-chip.test.ts
@@ -1,9 +1,9 @@
 import {assert, fixture} from '@open-wc/testing';
 import {assign, html} from 'element-vir';
 import {assertInstanceOf} from '../../element-testing/assertion-helpers';
-import {fixtureTest} from '../../element-testing/fixture-test';
+import {createFixtureTest} from '../../element-testing/fixture-test';
+import {assertIconEquals} from '../../element-testing/icon-test-helpers';
 import {queryThroughShadow} from '../../element-testing/query-through-shadow';
-import {assertIconEquals} from '../../element-testing/test-icon';
 import {Copy24Icon} from '../../icons';
 import {ToniqIcon} from '../toniq-icon/toniq-icon.element';
 import {ToniqChip} from './toniq-chip.element';
@@ -16,7 +16,7 @@ describe(ToniqChip.tagName, () => {
 
     it(
         'should render child elements',
-        fixtureTest(async () => {
+        createFixtureTest(async () => {
             const textToRender = 'Stuff in there';
             const iconToRender = Copy24Icon;
