+++ src/elements/toniq-toggle-button/toniq-toggle-button.test.ts
@@ -1,6 +1,6 @@
 import {assert, fixture} from '@open-wc/testing';
 import {assign, html} from 'element-vir';
-import {fixtureTest} from '../../element-testing/fixture-test';
+import {createFixtureTest} from '../../element-testing/fixture-test';
 import {getTextContentThroughShadow} from '../../element-testing/query-through-shadow';
 import {ToniqToggleButton} from './toniq-toggle-button.element';
 
@@ -12,7 +12,7 @@ describe(ToniqToggleButton.tagName, () => {
 
     it(
         'should render assigned text',
-        fixtureTest(async () => {
+        createFixtureTest(async () => {
             const textToRender = 'Buy Now';
 
             const rendered = await fixture(
